backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
crewai>=0.100.0
crewai-tools>=0.17.0
litellm>=1.50.0
matplotlib>=3.9.0
seaborn>=0.13.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pysqlite3-binary>=0.5.0
